% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idaweb-package.r
\docType{package}
\name{idaweb-package}
\alias{idaweb}
\alias{idaweb-package}
\title{idaweb: idaweb - Access Ground-Based Measurements provided by MeteoSwiss}
\description{
This package simplifies searching for and downloading ground-based measurements provided by MeteoSwiss through the Open Data API
}
