% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-data.r
\name{met_search}
\alias{met_search}
\title{Search Meteoswiss Data}
\usage{
met_search(
  from,
  to,
  tz = get_tzone(from, to),
  lon,
  lat,
  alt,
  abbr,
  name,
  canton,
  shortname,
  unit,
  group,
  description,
  granularity = c("T", "H", "D", "M", "Y"),
  language = c("en", "de", "fr", "it"),
  meta_data = idaweb::metadata,
  drop_nodata = TRUE
)
}
\arguments{
\item{from}{a character or POSIXt object. TODO}

\item{to}{a character or POSIXt object. TODO}

\item{tz}{character. TODO}

\item{lon}{character. TODO}

\item{lat}{character. TODO}

\item{alt}{numeric. TODO}

\item{abbr}{character. TODO}

\item{name}{character. TODO}

\item{canton}{character. TODO}

\item{shortname}{character. TODO}

\item{unit}{character. TODO}

\item{group}{character. TODO}

\item{description}{character. TODO}

\item{granularity}{character. TODO}

\item{language}{character. TODO}

\item{meta_data}{an object of class \code{met_metadata}. TODO}
}
\value{
object of class \code{met_metadata}. TODO
}
\description{
TODO
}
\examples{
 \dontrun{
 require(idaweb)
 # search for wind data
 meta_wind <- met_search('01.01.2025 to 04.02.2026', lon = '7.4..7.5', 
     lat = '46.9..47.3', granularity = c('d', 'm'), group = 'wind')
 # etc.
 }
}
