% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{print.met_assets}
\alias{print.met_assets}
\title{Print MeteoSwiss Assets}
\usage{
\method{print}{met_assets}(x, ...)
}
\arguments{
\item{x}{An object of class \code{met_assets}.}

\item{...}{Not used.}
}
\description{
Print MeteoSwiss Assets
}
