% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{print.met_collections}
\alias{print.met_collections}
\title{Print MeteoSwiss Collections}
\usage{
\method{print}{met_collections}(x, ...)
}
\arguments{
\item{x}{An object of class \code{met_collections}.}

\item{...}{Not used.}
}
\description{
Print MeteoSwiss Collections
}
