% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{print.met_datainventory}
\alias{print.met_datainventory}
\title{Print Data Inventory}
\usage{
\method{print}{met_datainventory}(x, ...)
}
\arguments{
\item{x}{An object of class \code{met_datainventory}.}

\item{...}{Further arguments passed to \code{print_dense}.}
}
\description{
Print Data Inventory
}
