% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{print.met_metadata}
\alias{print.met_metadata}
\title{Print MeteoSwiss Metadata}
\usage{
\method{print}{met_metadata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{met_metadata}.}

\item{...}{Not used.}
}
\description{
Print MeteoSwiss Metadata
}
