% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{print.met_parameters}
\alias{print.met_parameters}
\title{Print Parameter Metadata}
\usage{
\method{print}{met_parameters}(x, ...)
}
\arguments{
\item{x}{An object of class \code{met_parameters}.}

\item{...}{Further arguments passed to \code{print_dense}.}
}
\description{
Print Parameter Metadata
}
