% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{print.met_stations}
\alias{print.met_stations}
\title{Print Station Metadata}
\usage{
\method{print}{met_stations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{met_stations}.}

\item{...}{Further arguments passed to \code{print_dense}.}
}
\description{
Print Station Metadata
}
