% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-data.r
\name{search_by_datetime}
\alias{search_by_datetime}
\title{Filter Metadata by Date and Time}
\usage{
search_by_datetime(
  from,
  to,
  tz = get_tzone(from, to),
  meta_data = idaweb::metadata,
  drop_nodata = FALSE
)
}
\arguments{
\item{from}{Start date/time. See \code{\link{met_search}}.}

\item{to}{End date/time. See \code{\link{met_search}}.}

\item{tz}{Time zone for \code{from} and \code{to}. See \code{\link{met_search}}.}

\item{meta_data}{A \code{met_metadata} object or list thereof. Defaults to
\code{\link[idaweb]{metadata}}.}

\item{drop_nodata}{Logical. If \code{TRUE}, collections with an empty
inventory after filtering are removed.}
}
\value{
A \code{met_metadata} object or a list thereof.
}
\description{
Subsets a \code{met_metadata} object (or a list thereof) by temporal coverage.
Only station-parameter combinations whose data inventory overlaps with the
requested interval are retained.
}
\examples{
\dontrun{
meta <- search_by_datetime("01.01.2020", "31.12.2020")
meta <- search_by_datetime("01.01.2020", 
  meta_data = idaweb::metadata[[1]])
}
}
\seealso{
\code{\link{search_by_location}}, \code{\link{search_by_parameter}}
}
