% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.r
\name{collections}
\alias{collections}
\title{Fetch MeteoSwiss Open Data Collections (Ground-based) Measurements}
\usage{
collections(set_name = NULL)
}
\arguments{
\item{set_name}{character. TODO}
}
\value{
object of class \code{met_collections}. TODO
}
\description{
TODO
}
