% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.r
\name{datainventory}
\alias{datainventory}
\title{Access Meta Data Inventory Info}
\usage{
datainventory(meta_data, as_dt = FALSE, cols = NULL, uniq = !is.null(cols))
}
\arguments{
\item{meta_data}{\code{met_metadata}. TODO}

\item{as_dt}{logical. Default \code{FALSE}. TODO}

\item{cols}{character vector. Default \code{NULL}. TODO}

\item{uniq}{logical. Default \code{!is.null(cols)}. TODO}
}
\value{
object of class \code{met_datainventory}. TODO
}
\description{
TODO
}
